feat(list): expose listing component
ref: MANAGER-5369
