+++ packages/manager/modules/ng-layout-helpers/list/list-layout.component.js
@@ -1,0 +1,19 @@
+import controller from './list-layout.controller';
+import template from './template.html';
+import { componentBindings } from './list-layout.utils';
+
+export default {
+  bindings: {
+    ...componentBindings,
+    staticResources: '<?',
+    header: '@',
+    id: '@',
+    description: '<',
+    loadRow: '<?',
+    columns: '<',
+    customizableColumns: '<?',
+    defaultFilterProperty: '<',
+  },
+  controller,
+  template,
+};
