+++ packages/manager/modules/ng-layout-helpers/list/list-layout.controller.js
@@ -1,5 +1,5 @@
 import get from 'lodash/get';
-import set from 'lodash/set';
+import map from 'lodash/map';
 
 export default class ListLayoutCtrl {
   /* @ngInject */
@@ -9,6 +9,11 @@ export default class ListLayoutCtrl {
   }
 
   $onInit() {
+    this.datagridId = `dg-${this.id}`;
+    this.defaultFilterColumn = this.defaultFilterProperty;
+
+    this.getDisplayedColumns(this.columns);
+
     this.criteria = JSON.parse(this.filter).map((criteria) => ({
       property: get(criteria, 'field') || this.defaultFilterColumn,
       operator: get(criteria, 'comparator'),
@@ -21,12 +26,6 @@ export default class ListLayoutCtrl {
   }
 
   loadPage() {
-    const currentOffset = (this.paginationNumber - 1) * this.paginationSize;
-    set(
-      this.ouiDatagridService,
-      `datagrids.${this.datagridId}.paging.offset`,
-      currentOffset + 1,
-    );
     return this.$q.resolve({
       data: get(this.resources, 'data'),
       meta: {
@@ -39,8 +38,17 @@ export default class ListLayoutCtrl {
     return this.sort === property ? this.sortOrder.toLowerCase() : '';
   }
 
+  getDisplayedColumns(columns) {
+    this.displayedColumns = JSON.stringify(
+      map(
+        columns.filter(({ hidden }) => !hidden),
+        ({ name, property }) => name || property,
+      ),
+    );
+  }
+
   onPageChange({ pageSize, offset }) {
-    this.onListParamsChange({
+    this.onParamsChange({
       page: parseInt(offset / pageSize, 10) + 1,
       pageSize,
     });
@@ -53,15 +61,27 @@ export default class ListLayoutCtrl {
       reference: [criteria.value],
     }));
 
-    this.onListParamsChange({
+    this.onParamsChange({
       filter: JSON.stringify(filter),
     });
   }
 
   onSortChange({ name, order }) {
-    this.onListParamsChange({
+    this.onParamsChange({
       sort: name,
       sortOrder: order,
     });
+  }
+
+  onParamsChange(params) {
+    return this.onListParamsChange({
+      ...params,
+      columns: this.displayedColumns,
+    });
+  }
+
+  onColumnChange(id, columns) {
+    this.getDisplayedColumns(columns);
+    return this.onListParamsChange();
   }
 }
