+++ packages/manager/modules/ng-layout-helpers/list/template.html
@@ -1,0 +1,35 @@
+<oui-header class="mb-2" data-heading="{{:: $ctrl.header }}"></oui-header>
+<div data-ng-bind-html="$ctrl.description"></div>
+<oui-datagrid
+    data-ng-if="$ctrl.staticResources"
+    rows="$ctrl.resources"
+    id="{{ $ctrl.datagridId }}"
+    row-loader="$ctrl.loadRow($row)"
+    columns="$ctrl.columns"
+    customizable="$ctrl.customizableColumns"
+    criteria="$ctrl.criteria"
+    on-columns-parameters-change="$ctrl.onColumnChange(id, columns)"
+    on-page-change="$ctrl.onPageChange($pagination)"
+    on-criteria-change="$ctrl.onCriteriaChange($criteria)"
+    page-size="{{ $ctrl.paginationSize }}"
+    page="{{ $ctrl.paginationNumber }}"
+    on-sort-change="$ctrl.onSortChange($sort)"
+>
+</oui-datagrid>
+
+<oui-datagrid
+    data-ng-if="!$ctrl.staticResources"
+    id="{{ $ctrl.datagridId }}"
+    rows-loader="$ctrl.loadPage()"
+    row-loader="$ctrl.loadRow($row)"
+    columns="$ctrl.columns"
+    customizable="$ctrl.customizableColumns"
+    criteria="$ctrl.criteria"
+    on-columns-parameters-change="$ctrl.onColumnChange(id, columns)"
+    on-page-change="$ctrl.onPageChange($pagination)"
+    on-criteria-change="$ctrl.onCriteriaChange($criteria)"
+    page-size="{{ $ctrl.paginationSize }}"
+    page="{{ $ctrl.paginationNumber }}"
+    on-sort-change="$ctrl.onSortChange($sort)"
+>
+</oui-datagrid>
