+++ packages/manager/modules/ng-layout-helpers/package.json
@@ -16,6 +16,7 @@
   },
   "peerDependencies": {
     "@ovh-ux/ng-ovh-api-wrappers": "^4.0.6",
+    "@ovh-ux/ui-kit": "^4.2.0",
     "angular": "^1.7.9"
   }
 }
